Create helper script to derive innerHTML linter avoidance.
moz-build-helper.sh and its text files now seem to correctly produce the
current `lit.all.js` we've checked into the tree at `mozdist/lit.all.js`
(which is not currently checked into the tree.

Note that this somewhat obsoletes the contents of `moz-header.txt` but
it can now be easily changed, probably!